refactor: factor out pop_actions
