+++ execution-engine/contracts/hdac-system/pop/src/pop_impl/pop_actions/delegations_provider.rs
@@ -12,24 +12,11 @@ use types::{
     Key, U512,
 };
 
-pub struct ContractDelegations;
-pub struct Delegations(pub BTreeMap<DelegationKey, U512>);
-pub struct DelegationStat(pub BTreeMap<PublicKey, U512>);
+use super::delegations::{DelegationKey, DelegationStat, DelegationUnitForOrder, Delegations};
+use crate::pop_impl::ProofOfProfessionContract;
 
-#[derive(PartialOrd, Ord, PartialEq, Eq, Clone, Copy)]
-pub struct DelegationKey {
-    pub delegator: PublicKey,
-    pub validator: PublicKey,
-}
-
-#[derive(PartialOrd, Ord, PartialEq, Eq, Clone, Copy)]
-pub struct DelegationUnitForOrder {
-    pub validator: PublicKey,
-    pub amount: U512,
-}
-
-impl ContractDelegations {
-    pub fn read() -> Result<Delegations> {
+impl ProofOfProfessionContract {
+    pub fn read_delegations(&self) -> Result<Delegations> {
         let mut delegations = BTreeMap::new();
         for (name, _) in runtime::list_named_keys() {
             let mut split_name = name.split('_');
@@ -78,7 +65,7 @@ impl ContractDelegations {
     }
 
     /// Writes the current stakes to the contract's known urefs.
-    pub fn write(delegations: &Delegations) {
+    pub fn write_delegations(&self, delegations: &Delegations) {
         // Encode the stakes as a set of uref names.
         let mut new_urefs: BTreeSet<String> = delegations
             .0
@@ -111,7 +98,7 @@ impl ContractDelegations {
         }
     }
 
-    pub fn read_stat() -> Result<DelegationStat> {
+    pub fn read_delegation_stat(&self) -> Result<DelegationStat> {
         let mut delegation_stat = BTreeMap::new();
         for (name, _) in runtime::list_named_keys() {
             let mut split_name = name.split('_');
@@ -149,7 +136,7 @@ impl ContractDelegations {
         Ok(DelegationStat(delegation_stat))
     }
 
-    pub fn read_user_stat() -> Result<DelegationStat> {
+    pub fn read_delegation_user_stat(&self) -> Result<DelegationStat> {
         let mut delegation_stat = BTreeMap::new();
         for (name, _) in runtime::list_named_keys() {
             let mut split_name = name.split('_');
@@ -187,7 +174,8 @@ impl ContractDelegations {
         Ok(DelegationStat(delegation_stat))
     }
 
-    pub fn get_sorted_stat(
+    pub fn get_sorted_delegation_stat(
+        &self,
         delegation_stat: &DelegationStat,
     ) -> Result<Vec<DelegationUnitForOrder>> {
         let mut delegation_sorted: Vec<DelegationUnitForOrder> = Vec::new();
@@ -201,49 +189,5 @@ impl ContractDelegations {
         delegation_sorted.sort_by(|a, b| b.amount.cmp(&a.amount));
 
         Ok(delegation_sorted)
-    }
-}
-
-impl Delegations {
-    pub fn delegate(&mut self, delegator: &PublicKey, validator: &PublicKey, amount: U512) {
-        let key = DelegationKey {
-            delegator: *delegator,
-            validator: *validator,
-        };
-        self.0
-            .entry(key)
-            .and_modify(|x| *x += amount)
-            .or_insert(amount);
-    }
-
-    pub fn undelegate(
-        &mut self,
-        delegator: &PublicKey,
-        validator: &PublicKey,
-        maybe_amount: Option<U512>,
-    ) -> Result<U512> {
-        let key = DelegationKey {
-            delegator: *delegator,
-            validator: *validator,
-        };
-
-        match maybe_amount {
-            // undelegate all
-            None => self.0.remove(&key).ok_or(Error::NotDelegated),
-            Some(amount) => {
-                let delegation = self.0.get_mut(&key);
-                match delegation {
-                    Some(delegation) if *delegation > amount => {
-                        *delegation -= amount;
-                        Ok(amount)
-                    }
-                    Some(delegation) if *delegation == amount => {
-                        self.0.remove(&key).ok_or(Error::DelegationsNotFound)
-                    }
-                    Some(_) => Err(Error::UndelegateTooLarge),
-                    None => Err(Error::NotDelegated),
-                }
-            }
-        }
     }
 }
